order/requirements.txt
pydantic==1.10.12
pydantic[dotenv]
fastapi==0.95.1
uvicorn==0.17.6
psycopg2-binary==2.9.6
sqlalchemy
alembic==1.11.1
aio-pika
simplejson